修改购买页输入框位数过长的错误

See merge request !105

+++ hoomxb/Financing/ViewController/Buy/HXBFin_creditorChange_buy_ViewController.m
@@ -68,6 +68,7 @@ static NSString *const investString = @"立即投资";
     self.isColourGradientNavigationBar = true;
 //    [self getNewUserInfo];
     [self buildUI];
+    
     [self setUpDate];
     [self getBankCardLimit];
 }
@@ -283,13 +284,13 @@ static NSString *const investString = @"立即投资";
      _registerMultipleAmount     最低起投此金额多少倍
      _inputMoneyStr              输入的金额
      */
-    BOOL isHasContainsNonzeroDecimals = (int)([_inputMoneyStr doubleValue] * 100) % 100 != 0 ? true:false;//true:含非零小数
+    BOOL isHasContainsNonzeroDecimals = (long long)([_inputMoneyStr doubleValue] * 100) % 100 != 0 ? true:false;//true:含非零小数
     BOOL isMultipleOfMin = ((_inputMoneyStr.integerValue - _minRegisterAmount.integerValue) % _registerMultipleAmount.integerValue);//true表示非（最低倍数）的整数倍
     if (_inputMoneyStr.length <= 0) {
         [HxbHUDProgress showTextWithMessage:@"请输入投资金额"];
     }else{
         if (isHasContainsNonzeroDecimals) {
-            if ((int)([_inputMoneyStr doubleValue] * 100) == (int)([_availablePoint doubleValue] * 100)) {
+            if ((long long)([_inputMoneyStr doubleValue] * 100) == (long long)([_availablePoint doubleValue] * 100)) {
                 [self chooseBuyTypeWithSting:_btnLabelText];
                 return;
             }else{
